// apps/main/src/models/v2/controlplane/environmentModuleVersionPin.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */
import type { ModuleVersionPinStatus } from './moduleVersionPinStatus';

export interface EnvironmentModuleVersionPin {
  id: string;
  org_id: string;
  project_uuid: string;
  project_id: string;
  environment_uuid: string;
  environment_id: string;
  module_uuid: string;
  version_uuid: string;
  status: ModuleVersionPinStatus;
  /** @minimum 1 */
  resource_version: number;
  activation_event_id: string;
  bulk_operation_id?: string;
  override_operation_id?: string;
  override_target_version_uuid?: string;
  override_actor?: string;
  override_reason?: string;
  override_deployment_id?: string;
  created_by: string;
  created_at: string;
  updated_at: string;
  removed_at?: string;
}