// apps/main/src/models/v2/controlplane/moduleVersionPinBulkAction.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */

export type ModuleVersionPinBulkAction =
  (typeof ModuleVersionPinBulkAction)[keyof typeof ModuleVersionPinBulkAction];

export const ModuleVersionPinBulkAction = {
  pin: 'pin',
  unpin: 'unpin',
  discard: 'discard',
} as const;