// apps/main/src/models/v2/controlplane/moduleVersionPinBulkPreview.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */
import type { ModuleVersionPinBulkAction } from './moduleVersionPinBulkAction';
import type { ModuleVersionPinBulkPreviewItem } from './moduleVersionPinBulkPreviewItem';

export interface ModuleVersionPinBulkPreview {
  operation_id: string;
  fingerprint: string;
  action: ModuleVersionPinBulkAction;
  module_uuid: string;
  eligible: boolean;
  items: ModuleVersionPinBulkPreviewItem[];
}